server/requirements.txt
opuslib_next==1.1.4
pydub==0.25.1
audioop-lts==0.2.1
numpy==1.26.4
asyncudp==0.11.0
sounddevice==0.5.1
